pyproject.toml
[tool.poetry]
name = "configdb"
version = "3.0.5"
description = "Configuration Database"
authors = ["Observatory Control System Project <ocs@lco.global>"]
license = "GPL-3.0-only"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
django = "^4"
psycopg2-binary = ">=2.8,<2.9"
requests = ">=2,<3"
djangorestframework = "^3.13"
django-filter = "21.1"
django-reversion = "^5"
django-cors-headers = "^3.11"
ocs-authentication = "^0.2.1"
cerberus = "^1.3.4"
django-extensions = "^3.1.5"
ipython = "^8.1.1"
gunicorn = {version = "^20.1.0", extras = ["gevent"]}
mixer = "^7.2.1"
PyYAML = "^6.0"
time_intervals = ">=1.0,<2.0"
uritemplate = "^4.1.1"

[tool.poetry.dev-dependencies]
coverage = "^6.3.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
